add regex support for error codes
Some error codes contain wildcards and would apply to a range
of error codes. To be able to support wildcard'ed error codes
I'm allowing developer to define a regex as an ID.

When a specific error code exists but it also matches one of the regexs error codes then all found Messages are logged. This allows
 to define generic messages versus specific messages once more is
 known about the specific code.

Refactored some small things like line highlighting and the logging
of the playstation error code. The line highlighting would occasionally not work, the current approach always works.

Data comment;
The change is transparent and still works with older versions of
the app. The newly added codes which have a regex will simply not
be matched as the original code does a simple equal compare.